requirements.txt
scrapy
spacy
langdetect
pytesseract
pdf2image
Pillow